Bug fixes
Fixed display errors for the warning checks, also realized that I was adding the temperature buffer in the pump controller when I should've been subtracting it, so I fixed that too. Code has now been tested on the actual system with thermocouples+flow sensor plugged in, seems to work as intended after the fixes.